adding a user model

+++ public/homeThree.js
@@ -1,26 +1,134 @@
 
-function clickedBtn(){
-    console.log('button clicked')
-}
-
 window.onload = () => {
 
+    setEventListeners()
 
-    //clickMe.onclick = clickedBtn;
+    const editDivs = document.getElementsByClassName('editDivs');
 
-    const getButtons = document.getElementsByClassName('getMovie');
+            for (const div of editDivs){
 
-    for (const button of getButtons) {
-        button.onclick = reqSingleMovieData
-    };
+                div.style.display = 'none'
 
-    const deleteButton = document.getElementsByClassName('deleteMovie');
+            };
 
-    for (const btn of deleteButton) {
-        btn.onclick = deleteSingleMovie
-    };
+    const displayDivs = document.getElementsByClassName('displayDivs');
+                
+            for (const div of displayDivs){
+
+            div.style.display = 'flex'
+
+            };
+
+    //clickMe.onclick = clickedBtn;  
+
+    allMovies.style.display = 'flex';
+
 
 }
+
+function setEventListeners() {
+    const getMovie = document.getElementsByClassName('getMovie');
+
+            for (const button of getMovie) {
+                 button.onclick = reqSingleMovieData
+            };
+
+    const editMovieBtn = document.getElementsByClassName('editMovieBtn');
+
+            for (const button of editMovieBtn) {
+                 button.onclick = changeEditView
+            };
+
+    const editSubmitButtons = document.getElementsByClassName('submitEdit');
+
+            for (const button of editSubmitButtons){
+
+                button.onclick = submitEditReq
+
+            };
+
+    
+}
+
+
+function submitEditReq() {
+
+    console.log('asdf');   
+    
+
+    const movieId = this.parentElement.parentElement.id;
+    const form = this.parentElement.childNodes[0];
+
+    let validationErr = [];
+
+    let reqBody = {};
+
+    
+
+        for (const input of form) {
+
+            let inputValue = input.value.trim();
+
+            if (inputValue != '') {
+
+                reqBody[input.name] = inputValue;
+
+            }
+
+
+           if (input.validationMessage != ''){
+               
+            validationErr.push(`${input.name}: ${input.validationMessage}`);
+
+            }
+
+            // if ( !(new RegExp(/imdb/).test(form.imdb_link)) )  {
+
+            //         validationErr.push('IMDB Link provided was not valid')
+
+            // }
+
+
+
+           if (validationErr.length > 0) {
+
+            const message = `Edit not made \n\n ${validationErr.join('\n')}`
+
+            return alert(message)
+
+           }
+
+           
+
+        }
+
+        console.log(reqBody)
+
+        const endpoint = `${location.origin}/movie/patch/${movieId}`
+        const reqObj = {
+            method: 'PATCH',
+            body: JSON.stringify(reqBody),
+            headers: {
+             'Access-Control-Allow-Origin': '*',
+             Accept: 'Application/json',
+             'content-type': 'application/json'
+            }
+            
+
+        };
+
+        fetch(endpoint, reqObj)
+            .then(rs => {return rs.json(})
+            .then(res => {console.log(res)})
+            .catch( err => {console.log(err) });
+
+         //check if successful
+             //if succesful switch back to original screen
+             //if not successful alert user of error with message
+
+     
+}
+
 
 function reqSingleMovieData() {
     const movieId = this.parentElement.id;
@@ -44,23 +152,59 @@ function reqSingleMovieData() {
     // xhr.send
 }
 
+
 function deleteSingleMovie() {
 
+    let confirm = prompt('Type "confirm" to delete document');
 
-    const movieId = this.parentElement.id;
+    if (confirm != 'confirm') {
+        
+        return alert('Document deletion canceled')
+
+    }
+
+
+    const movieId = this.parentElement.parentElement.id;
 
     const endpoint = `${location.origin}/movie/delete/${movieId}`;
 
     reqObj = {
-        method: 'DELETE'
+         method: 'DELETE'
     };
 
     fetch(endpoint, reqObj)
-    .then(rs => {return rs.json()})
-    .then(res => {
-        console.log(res)
-    })
+    .then(rs => {
+    
+    if (rs.status === 200) { //doc DB deletion success, delete element from FE.
 
+        this.parentElement.parentElement.remove();
 
-
+        } 
+    
+        }    
+    )
 }
+
+
+function changeEditView() {
+
+    const movieDivChildren = this.parentElement.parentElement.childNodes;
+
+    movieDivChildren.forEach( node => {
+
+        if (node.className === 'displayMovie'  ||  node.className === 'editMovie') {
+
+            node.style.dispay = node.style.display == 'none' ? 'flex' : 'none';
+        
+
+        }
+
+    })
+}
+
+
+
+
+
+
+
